rename funds.tsx to FundsPage.tsx
